Cleans up documentation and fixes a bug.
This commit simplifies the example in the documenation. It also fixes
a bug that would prevent a valid authentication from succeeding due to
a verification against an invalid "embedded key". I'm not entirely happy
with the final check for a valid token (by simply checking the username)
but it will do.