Refactoring and extra debugging
Refactored the decryption method in util.Crypto to accept a
keystore.Key as a parameter (makes more sense). Also refactored the
`decryptData` method of token.Token to use the decryption method in
util.Crypto (instead of repeating the same code in multiple places).
Finally, some extra debug logging was aded that I needed to solve
an implementation problem.